(refactor) Create an aggregator file to all action creators
